Create setup.sh
This setup file is designed so that after the git is cloned to a Kali Linux instance, the setup file will resolve any necessary libraries.  This will be extremely useful for when using Kali as a boot disk or live instance.  You can easily have these scripts ready to run as needed each time you initiate the Kali Linux OS.