Create username_generator.py
This file takes the Top 1000 surname list from the US Census and produces a list of usernames by prepending every letter of the alphabet to the surname.  This script also supports the merging of the output password list with another list, which it supports by appending the data or prepending the data.